Bound live Skill resource reads by actual bytes
